+++ src/objects/impls.rs
@@ -5,6 +5,7 @@ mod deposit;
 mod flag;
 mod mineral;
 mod nuke;
+mod power_creep;
 mod resource;
 mod room;
 mod room_terrain;
@@ -28,9 +29,9 @@ mod structure_tower;
 mod tombstone;
 
 pub use self::{
-    creep::{Bodypart, MoveToOptions},
+    creep::Bodypart,
     room::{
-        AttackEvent, AttackType, BuildEvent, Event, EventType, ExitEvent, FindOptions,
+        AttackEvent, AttackType, BuildEvent, Effect, Event, EventType, ExitEvent, FindOptions,
         HarvestEvent, HealEvent, HealType, LookResult, ObjectDestroyedEvent, Path,
         PositionedLookResult, RepairEvent, ReserveControllerEvent, Step, UpgradeControllerEvent,
     },