+++ src/objects/impls/room.rs
@@ -13,8 +13,8 @@ use stdweb::{Reference, Value};
 
 use crate::{
     constants::{
-        Color, Direction, ExitDirection, FindConstant, Look, LookConstant, PowerType, ResourceType,
-        ReturnCode, StructureType, Terrain,
+        Color, Direction, EffectType, ExitDirection, FindConstant, Look, LookConstant, PowerType,
+        ResourceType, ReturnCode, StructureType, Terrain,
     },
     local::{Position, RoomName},
     memory::MemoryReference,
@@ -73,7 +73,11 @@ impl Room {
     {
         let pos = at.pos();
         js_unwrap!(@{self.as_ref()}.createConstructionSite(
-            pos_from_packed(@{pos.packed_repr()}),
+            // pos_from_packed(@{pos.packed_repr()}),
+            // workaround - passing with a position and a name
+            // currently broken, use x,y instead
+            @{pos.x()},
+            @{pos.y()},
             __structure_type_num_to_str(@{ty as u32}),
             @{name}
         ))
@@ -238,12 +242,12 @@ impl Room {
         // See https://docs.rs/scoped-tls/0.1/scoped_tls/
         COST_CALLBACK.set(&callback_lifetime_erased, || {
             let v = js! {
-                return @{&self.as_ref()}.search(
+                return @{&self.as_ref()}.findPath(
                     pos_from_packed(@{from.packed_repr()}),
                     pos_from_packed(@{to.packed_repr()}),
                     {
                         ignoreCreeps: @{ignore_creeps},
-                        ignoreDestructibleStructures: @{ignore_destructible_structures}
+                        ignoreDestructibleStructures: @{ignore_destructible_structures},
                         costCallback: @{callback},
                         maxOps: @{max_ops},
                         heuristicWeight: @{heuristic_weight},
@@ -759,6 +763,15 @@ pub struct PowerEvent {
     pub target_id: String,
     pub power: PowerType,
 }
+
+#[derive(Clone, Debug, PartialEq, Eq, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Effect {
+    pub effect: EffectType,
+    pub level: Option<u8>,
+    pub ticks_remaining: u32,
+}
+js_deserializable! {Effect}
 
 pub enum LookResult {
     Creep(Creep),