+++ src/objects/impls/structure_tower.rs
@@ -1,18 +1,27 @@
 use crate::{
     constants::ReturnCode,
-    objects::{Creep, Structure, StructureTower},
+    objects::{Attackable, SharedCreepProperties, StructureProperties, StructureTower},
 };
 
 impl StructureTower {
-    pub fn attack(&self, target: &Creep) -> ReturnCode {
+    pub fn attack<T>(&self, target: &T) -> ReturnCode
+    where
+        T: Attackable,
+    {
         js_unwrap! { @{self.as_ref()}.attack( @{target.as_ref()} ) }
     }
 
-    pub fn heal(&self, target: &Creep) -> ReturnCode {
+    pub fn heal<T>(&self, target: &T) -> ReturnCode
+    where
+        T: SharedCreepProperties,
+    {
         js_unwrap! { @{self.as_ref()}.heal( @{target.as_ref()} ) }
     }
 
-    pub fn repair(&self, target: &Structure) -> ReturnCode {
+    pub fn repair<T>(&self, target: &T) -> ReturnCode
+    where
+        T: StructureProperties,
+    {
         js_unwrap! { @{self.as_ref()}.repair( @{target.as_ref()} ) }
     }
 }